+++ crate/event_loop/src/event_loop/common/rate_limiter.rs
@@ -1,0 +1,6 @@
+use governor::{
+    clock::DefaultClock,
+    state::{direct::NotKeyed, InMemoryState},
+};
+
+pub(crate) type RateLimiter = governor::RateLimiter<NotKeyed, InMemoryState, DefaultClock>;