+++ crate/workspace_tests/src/nginee_event_loop.rs
@@ -1,7 +1,0 @@
-#[cfg(feature = "rate_limit")]
-mod error;
-mod event_handling_outcome;
-#[cfg(not(target_arch = "wasm32"))]
-mod event_loop;
-#[cfg(feature = "rate_limit")]
-mod rate_limit;