+++ examples/event_loop/src/main.rs
@@ -1,4 +1,4 @@
-use example_event_loop::{run, Error};
+use event_loop::{run, Error};
 
 fn main() -> Result<(), Error> {
     run()