+++ examples/event_loop_console/main.rs
@@ -1,5 +1,0 @@
-use example_event_loop_console::{run, Error};
-
-fn main() -> Result<(), Error> {
-    run()
-}