+++ examples/event_loop_rate_limit/src/lib.rs
@@ -79,10 +79,10 @@ pub fn renderer(count: Arc<Mutex<u32>>) -> EventHandler<Error> {
     })
 }
 
-#[cfg(not(target_arch = "wasm32"))]
+#[cfg(all(not(target_arch = "wasm32"), not(feature = "window")))]
 type ReturnValue = Result<(), Error>;
 
-#[cfg(target_arch = "wasm32")]
+#[cfg(any(target_arch = "wasm32", feature = "window"))]
 type ReturnValue = ();
 
 /// Runs the application.