+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 * `event_loop` crate provides an application event loop. ([#2])
 * `"rate_limit"` feature allows event loop handlers to be rate limited. ([#5])
 * Contribution guide. ([#7])
+* `"window"` feature runs the event loop within `winit`'s event loop. ([#9])
 
 [#2]: https://github.com/azriel91/nginee/issues/2
 [#5]: https://github.com/azriel91/nginee/issues/5
 [#7]: https://github.com/azriel91/nginee/issues/7
+[#9]: https://github.com/azriel91/nginee/issues/9